+++ my-app/src/views/admin/dashboard/dashboard.styles.css
@@ -1,8 +1,8 @@
-.admin-dashboard {
+/* .admin-dashboard {
   display: flex;
   height: 100vh;
   background-color: #333333;
-  padding: 0; /* O ajustar el padding si es necesario */
+  padding: 0; 
   overflow: hidden;
 }
 
@@ -15,7 +15,7 @@
   align-items: flex-start;
   border-radius: 8px;
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
-  margin-left: -20px; /* Ajusta este valor según sea necesario */
+  margin-left: -20px; 
 }
 
 .sidebar h2 {
@@ -42,10 +42,10 @@
   background-color: rgba(255, 255, 255, 0.047); 
 }
 .dashboard-display{
-  display: flex; /* Utiliza flexbox para alinear los elementos */
-  justify-content: center; /* Centra horizontalmente */
-  align-items: center; /* Centra verticalmente */
-  height: 100%; /* Asegura que el contenedor ocupe todo el alto disponible */
+  display: flex; 
+  justify-content: center; 
+  align-items: center; 
+  height: 100%; 
   color: #ffffff;
   margin: auto;
 }
@@ -56,7 +56,7 @@
   background-color: rgba(255, 255, 255, 0.1); 
   border-radius: 8px;
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.3); 
-  overflow: auto; /* Esto permitirá que el contenido tenga scroll si se desborda */
+  overflow: auto; 
   max-width: 100%;
  
 }
@@ -76,8 +76,8 @@
 .manage-requests ul {
   list-style-type: none;
   padding: 0;
-  max-height: 600px; /* Ajusta esta altura según tus necesidades */
-  overflow-y: auto; /* Habilita el scroll vertical */
+  max-height: 600px; 
+  overflow-y: auto; 
   scrollbar-width: thin;
 }
 
@@ -94,7 +94,7 @@
 }
 .manage-pets li:last-child,
 .manage-requests li:last-child {
-  margin-bottom: 0; /* Elimina el margen inferior del último elemento */
+  margin-bottom: 0; 
 }
 .manage-pets li:hover,
 .manage-requests li:hover {
@@ -148,4 +148,48 @@
 .delete-button:hover {
   background-color: rgba(255, 255, 255, 0.2); 
 }
+ */
 
+ .admin-dashboard {
+  display: flex;
+  height: 100vh;
+  background-color: #333333;
+}
+
+.sidebar {
+  width: 250px;
+  background-color: rgba(0, 0, 0, 0.359);
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  border-radius: 8px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
+  position: sticky;
+  top: 0;
+  height: 100vh;
+}
+
+.sidebar h2 {
+  font-size: 1.2rem;
+  margin-bottom: 20px;
+  color: #fdfcfc; 
+}
+
+.sidebar-button {
+  width: 100%;
+  padding: 10px 15px;
+  font-size: 0.9rem;
+  background-color: rgba(255, 255, 255, 0.1); 
+  color: #ffffff;
+  border: 1px solid rgba(255, 255, 255, 0.3); 
+  border-radius: 4px;
+  text-align: left;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+  margin-bottom: 10px;
+}
+
+.sidebar-button:hover {
+  background-color: rgba(255, 255, 255, 0.047); 
+}
